// SKYNET server/GC/730/generated/cs2.ts
export interface ProtoDescriptor<TMessage> {
    readonly name: string;
}

export interface GcRoute<TRequest, TResponse> {
    readonly requestId: number;
    readonly request: ProtoDescriptor<TRequest>;
    readonly responseId: number;
    readonly response: ProtoDescriptor<TResponse>;
}